+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: Option to use an existing suite object in the suite command
+
 ## 1.5.2 - 2020-05-07
 
 * Fixed: Remove the Hide parameter for the second Pester It call too